Feature: Implement Mechanical and Atmospheric Synthesis for Planetary Resonance
- Added 10% Frequency Bonus in Gematria for cultural resonance
- Implemented timeline-specific coherence drain rates (Ancient: 0.5, Volatile: 2.0)
- Integrated atmosphericColor into HUD separators and scanning labels
- Updated Player.adjustCoherence to support fine-grained double delta
- Verified resonance bonuses and UI color shifts through manual vibe checks

+++ src/main/groovy/com/endlesstransit/core/Player.groovy
@@ -19,8 +19,8 @@ class Player {
         coherence = 100
     }
 
-    void adjustCoherence(int delta) {
-        coherence = Math.min(maxCoherence, Math.max(0, coherence + delta))
+    void adjustCoherence(double delta) {
+        coherence = Math.min(maxCoherence, Math.max(0, (coherence + delta) as int))
     }
 
     void listInventory() {
